+++ contracts/Core/parameters/interfaces/ICollectionManagerParams.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface ICollectionManagerParams {
-    function setMaxTolerance(uint16 _maxTolerance) external;
+    function setMaxTolerance(uint32 _maxTolerance) external;
 }
