+++ contracts/Core/parameters/interfaces/IRewardManagerParams.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 interface IRewardManagerParams {
-    function setPenaltyNotRevealNum(uint16 _penaltyNotRevealNumerator) external;
+    function setPenaltyNotRevealNum(uint32 _penaltyNotRevealNumerator) external;
 
     function setBlockReward(uint256 _blockReward) external;
 
@@ -10,5 +10,5 @@ interface IRewardManagerParams {
 
     function setMaxAge(uint32 _maxAge) external;
 
-    function setMaxTolerance(uint16 _maxTolerance) external;
+    function setMaxTolerance(uint32 _maxTolerance) external;
 }
